chore(MessageBar): Refactor types in MessageBar so we don't have to pass optional types around
Previously we were passing in a lot of optional types required for the AttachMenu, which is confusing. This allows for type detection. This could probably use more refactoring to figure out a way to duplicate less, but I am on a time crunch.
