+++ packages/module/src/AttachMenu/AttachMenu.tsx
@@ -22,17 +22,17 @@ export interface AttachMenuProps extends DropdownProps {
   /** Items in menu */
   filteredItems: React.ReactNode;
   /** A callback for when the input value changes. */
-  handleTextInputChange?: (value: string) => void;
+  handleTextInputChange: (value: string) => void;
   /** Flag to indicate if menu is opened. */
-  isOpen?: boolean;
+  isOpen: boolean;
   /** Reference to the menu  */
-  menuRef?: React.RefObject<any>;
+  menuRef: React.RefObject<any>;
   /** Additional properties to pass to the Popper */
   popperProps?: ExtendedDropdownPopperProps;
   /** Callback to change the open state of the menu. Triggered by clicking outside of the menu. */
   onOpenChange: (isOpen: boolean) => void;
   /** Keys that trigger onOpenChange, defaults to tab and escape. It is highly recommended to include Escape in the array, while Tab may be omitted if the menu contains non-menu items that are focusable. */
-  onOpenChangeKeys: string[];
+  onOpenChangeKeys?: string[];
   /** Function callback called when user selects item. */
   onSelect?: (event?: React.MouseEvent<Element, MouseEvent>, value?: string | number) => void;
   /** Placeholder for search input */
@@ -70,7 +70,7 @@ export const AttachMenu: React.FunctionComponent<AttachMenuProps> = ({
       <MenuSearchInput>
         <SearchInput
           aria-label="Filter menu items"
-          onChange={(_event, value) => handleTextInputChange && handleTextInputChange(value)}
+          onChange={(_event, value) => handleTextInputChange(value)}
           placeholder={searchInputPlaceholder}
         />
       </MenuSearchInput>
